requirements.txt
annotated-types==0.7.0
anyio==4.6.0
asyncmy==0.2.9
certifi==2024.8.30
cffi==1.17.1
click==8.1.7
cryptography==43.0.1
distlib==0.3.9
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.112.4
fastapi-cli==0.0.5
filelock==3.16.1
h11==0.14.0
httpcore==1.0.6
httptools==0.6.1
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jwcrypto==1.5.6
markdown-it-py==3.0.0
MarkupSafe==3.0.1
mdurl==0.1.2
mysql-connector==2.2.9
mysql-connector-python==9.0.0
orjson==3.10.7
packaging==24.1
passlib==1.7.4
platformdirs==4.3.6
pluggy==1.5.0
pycparser==2.22
pydantic==2.8.2
pydantic-extra-types==2.9.0
pydantic-settings==2.5.2
pydantic_core==2.20.1
Pygments==2.18.0
PyJWT==2.9.0
PyMySQL==1.1.1
pytest==8.3.3
pytest-asyncio==0.24.0
python-dotenv==1.0.1
python-multipart==0.0.12
PyYAML==6.0.2
rich==13.9.2
router==0.1
shellingham==1.5.4
sniffio==1.3.1
starlette==0.38.6
typer==0.12.5
typing_extensions==4.12.2
ujson==5.10.0
uvicorn==0.31.1
uvloop==0.20.0
virtualenv==20.27.0
watchfiles==0.24.0
websockets==13.1
